+++ noxfile.py
@@ -49,7 +49,6 @@ HYPOTHESIS_BINARY_ONLY = ("--only-binary", "hypothesis")
         "3.13t",
         "3.14t",
         "3.15t",
-        "pypy3.9",
         "pypy3.10",
         "pypy3.11",
     ],
@@ -675,7 +674,7 @@ def _bump(session: nox.Session, *, version: str, file: Path, kind: str) -> None:
 @contextlib.contextmanager
 def _replace_file(
     original_path: Path,
-) -> Generator[tuple[IO[str], IO[str]], None, None]:
+) -> Generator[tuple[IO[str], IO[str]]]:
     # Create a temporary file.
     fh, replacement_path = tempfile.mkstemp()